Double URL encode base64 encoded displayURL
As best we can tell right now, DFP URL decodes our string once in transit.
This breaks delivery of the displayUrl string to the line item because '=' is
an illegal character in line item macro values.  The solution is to URL
encode the base64 buffer twice, so DFP only encounters the sequence %3D.

Change-Id: Ia3f1203ca2f7f148bfd71c3a657d9bf703e877b7
JIRA: EE-96

Former-commit-id: d04f860143b98d43105b37d1a6109a622b8e4e11