Fix race in NetworkManagerTests
The NetworkManger's callbacks to its delegate are always
dispatched to the main thread, but callbacks to the response
handler happen directly in the thread that the dataTask's
completion handler runs on (usually background).

If we're going to verify that the delegate callbacks were
made, we need to do it by async dispatch to the main thread
to guarantee that the validation check runs after the
delegate is called.

Change-Id: I26045f6b236ae969c7584e917c3024ad34a14380

Former-commit-id: a8203806475fecb40241f577e8c1eb37956177a8