Adds the CRBidToken datatype.
The CRBidToken opaquely encapsulates a UUID and implements NSCoding
so that it may be used as a key in NSDictionary. The UUID is used
to ensure randomness and correct implementation of the hash and
isEqual: methods. Instances of this type will be used as keys in
BidToken -> TokenValue dictionaries.

Change-Id: I6edfb1ca0c015502591cc3a23d745f3159e3646d
JIRA: EE-269

Former-commit-id: a5358eb2e70804c359c7de84cc42eca2a8157845