Adds CR_TokenCache
CR_TokenCache maps the tokens to the creatives so that the publisher can easily request a creative for ad using a token.
CR_TokenCache can also generate a randomized token for each bid, making sure each token only maps to a certain creative, enabling Criteo to make sure the correct creative is displayed by the publisher.
Tests for CR_TokenCache are added, which mainly test the two methods working correctly - 1. requesting a unique token for a bid and 2. requesting the correct creative with a token.
For the test, we need to compare Token Values, so a method isEqual for CR_TokenValue is also added.

Change-Id: Icaae34a8e1b9c29347fe3a54c9b9621601e329c3
JIRA: EE-269

Former-commit-id: 33bf265341a23e3a057eaad03152118c5a586123