Fix to remove initialization of CR_DeviceInfo in cacheAdUnit conversion
Initializing CR_DeviceInfo is about ~2 MB as it includes initializing webViews
and userAgent. Every call to cacheAdUnitToAdUnit was creating a new
instance of the CR_DeviceInfo class as screenSize was an instance variable of this class.
Converting getScreenSize to a class method to avoid initialization of this class
unnecessarily.

Change-Id: Ic76e71634b9cc97884c2f9b873bdf04d06f0521f
JIRA:EE-339

Former-commit-id: bd04d1a006b50cf64f2280c5e5e71bdb510c8f87