Persist the Kill Switch
Make the config kill switch persistent from one run of an app using the
SDK to another. Note that the bug of EE-413 results from the fact that
bids can be fetched and ads displayed before the call to the config
endpoint returns with the kill switch value. Persisting the kill switch
doesn't solve the issue the first time the app is run after the kill
switch is set, but does guarantee it on subsequent runs (i.e. "eventual
correctness"). This is adequate to solve the issue.

Change-Id: I066dadef2835b437a25aafd233c4f961dc7f24f6
JIRA: EE-413

Former-commit-id: 301259cc14549d33fc0f6cf8a8130739f989f04c