Prevent double fetch from CDB for same adUnit
Added new dictionary bidFetchTrackerCache to track adUnits for which bids are
being fetched from CDB. The cache is managed by the networkManager and is set
with an adunit before a CDB invocation is done for it in the API handler. It is
unset when the network call completes - with or without failure. Before making an invocation, the
cache is checked for the adUnit and if found, we return.

Change-Id: I2059780a8bba56dbefd720ac5357c46c4e9fa557
JIRA:EE-464

Former-commit-id: 1f20e0fcbed0c9e4087b41b9c0b0e05da224b583