EE-473 refactor MoPub mediation adapter to hide internal properties and methods.
Moved declarations from CRBannerCustomEvent+Internal.h and
CRInterstitialCustomEvent+Internal.h so that private properties
are declared in the .m files for the two classes, and test-only initializers
are declared in the test files themselves (which are not available to the end
users). This minimizes the visibility of private properties and test-only
methods. It does mean that private properties needed for testing are re-declared
in the test code, but this is a minor annoyance, and I consider it a better
solution than having a separate .h file for the private properties, which is
what we're moving away from anyway. The two +Internal.h files are deleted.

Source files (seen by end users):
deleted:    CRBannerCustomEvent+Internal.h
deleted:    CRInterstitialCustomEvent+Internal.h
modified:   CRBannerCustomEvent.m
modified:   CRInterstitialCustomEvent.m

Test files:
modified:   CRBannerCustomEventTests.m
modified:   CRInterstitialCustomEventTests.m

Other:
modified:   project.pbxproj

Change-Id: I546f8a871a31048dab46ffa800f952f24565fddf
JIRA: EE-473