EE-415 Allow CRInterstitial to be reused for multiple ads.
The problem was in the CR_InterstitialViewController class. A new WKWebView is created
each time a new ad is loaded, either by loadAd or loadAdWithBidToken. However, the
web view is only added to the view hierarchy in viewDidLoad, which is only called
once for the lifetime of the view controller. This was solved by moving that code,
and other code in viewDidLoad, to viewDidAppear, which is called each time the web
view is displayed.

Also, to make sure that subsequent ads aren't dismissed early, the GCD block that
dismisses the ad after 7 seconds must be canceled if the user dismisses the ad.

modified: CR_InterstitialViewController.h
modified: CR_InterstitialViewController.m
modified: CRInterstitialDelegateTests.m
modified: CRInterstitialTests.m
modified: CRInterstitialViewControllerTests.m

Change-Id: Ie1a6fed6a1e91ea73a4f9a6d08c3445c901fd618
JIRA: EE-415

Former-commit-id: 22fe74ac526ec63fa09ee98ca9fab8e40fc17453