EE-494 2:Set isNative as true in CdbRequest
isNative key value in CdbRequest json is set to true if isNative is true in CR_CacheAdUnit.
Also, when the bidResponse contains nativeAssets, the bidCache key is constructed with isNative
as true. As EE-492 is not yet implemented, nativeAssets has been declared as a string in CdbBid object.
Unit tests.

Change-Id: Id2989a6ce326e008ea80e75b0cb0482c03800b5b
JIRA:EE-494

Former-commit-id: dd0cb06519a98ff86a349eb2830edb8b0cffadd8