Add header bidding section for test app
Added header bidding sections for DFP and MoPub. Included a custom native section in DFP.
Included sections for Banner 320x50, 300x250 and Interstitial. Also, included the section
for logging network requests that was already present in the test app in this repo. For the scope
of this ticket, only added header bidding sections.

Change-Id: Ib15d89d394287c8e9df1eac803e8aff44a597f7e
JIRA:EE-504