EE-355 Batch 1 changes
Modified the CR_TokenValue class to have a CRAdUnit object as a property
instead of just a CRAdUnitType, an enum that is part of the
CRAdUnit. This allows us to store the adUnitId as well as adUnitType.
The adUnitId is needed to fix EE-355.

Also fixed the copyWithZone method of CRBidToken, which was returning
'self'. A copy should always return a new object.

Change-Id: Ie404ba963fa0d272cbf02221969342bfe13308fe
JIRA: EE-355

Former-commit-id: b1d2808b131d559286b7e9a378d35880201fabfb