Add a target SDKUnitTests attached it to the scheme CriteoPublisherSdk
SDKUnitTests contains only tests that are under 200ms and without multithreading. The goal is to be fast and determinist. The other tests can be added later in an SDKIntegrationTests for example.  SDKUnitTests one can be used by developper for iterating quickly and SDKUnitTests + SDKIntegrationTests  can be run in the presubmit.

The pubsdkTest is still available and works. It remains like that for having a smooth transition on the CI.

Change-Id: Iaa8b49c3a33cadf844215bacf041254881fae1cc
JIRA: None

Former-commit-id: 4fb51d6ab71dbb9cbb167588b94f6a537b1fe0e0