Allow to create many Criteo instances for the tests
The Criteo class MUST be a shared instance on the public API.
The Criteo class SHOULD not be used as a shared instance in tests. For now,
some tests classes use the swizzling of OCMock and so mocking class that are
actually tested. It is error prone, may implies internal shared instances across the tests
(no isolation) and we want to be able to simply inject classes.

Change-Id: I0dec11d79296b826e9019edb3fb4421a7c0f7e85
JIRA: EE-593

Former-commit-id: 9f7640ac5be1137ccf7aa47a316fd14986809b62