Use NSNumber instead of BOOL in inner classes of GDPR
The goal is have the YES|NO|nil values available so that we are
able to know if the value has been set or not. It will help us
to not set a default value in the bid request if the value doesn't
exist or is invalid.

In objective-c there is no NSBoolean but we use NSNumber instead.

We first do this modification in the inner classes of CR_GDPR to have
less modifications in the next commits that will modify the CR_GDPR
class and its usages in CR_APIHandler.

Change-Id: I6c7f8b4531d6bca2e47d69a0fb51fb9a65868dfa
JIRA: EE-977

Former-commit-id: 8038a47d89c5c95f3f5fef05a6ae725862f4b84a