Extract all the serializations related to bid request in dedicated classes
Leitmotiv: While adjusting the serialization of the bid request for accepting
the new API of CR_GDPR, I assess that CR_APIHandler has too
many responsibilities and becomes entropic with new entities like the
impression ID and the CR_CdbRequest. I decide to extract one
responsibility from it.

This refactoring is backed by unit tests and integration tests
related to the CR_APIHandler. The next commit will move the
tests related to the serialization to the appropriated testing classes.

Change-Id: I6b5d4d500eb3c7114ac7361fb0cb59c6fd469188
JIRA: EE-977

Former-commit-id: 123a4cd6cbd2dd103a8d4150ff6427f7844b86ad