Call the completion of a Bid request even with empty response
After a discussion with Pierre, we agree that it is absurd
that in some cases, the completion handler is not called.
Initially I thought about return an error but Pierre suggested
that sometime we would have an empy data (a.k.a no bid) and it
is not necessary an error. So we simply call the handler with a
nil value. The CR_BidManager handles already this case.

Change-Id: Ib11efa5e59c6579067a450467489ba10962c00a3
JIRA: EE-970

Former-commit-id: 214a29ce0451b68a4f58cf9816235d14a3fdebfa