Don't update the feedback of an empty bid when getting this last one
The new code in BidManager changes the cache management:

BEFORE, when we got an empty bid, the "if (bid.isExpired)" was
positive. So we created a new feedback message because
setExpiredForImpressionId:bid.impressionId was called and we
remove the empty bid from the cache.

AFTER, when we get an empty bid, we keep it in the cache
and we don't create a new feedback message.
As Pierre suggested, it is awkward to keep empty bids in
the cache. It can be interesting to not add empty bids
when the bid request doesn't return a bid or fail for
a given adunit.

Change-Id: Id64282d73818a5401f7dfbe427130d9afc273c29
JIRA: EE-950

Former-commit-id: 46921aaef01d84d57610e11017c9c9ada0680812