Extract CSM logic outside bid manager
The bid manager already have a lot of concerns. And CSM feature may be
extended. For instance, a feature flag should be added, and CSM IO
should be done on worker threads.
To avoid adding more and more concerns in the bid manager, the CSM logic
is extracted and bid manager only use a high level controller.

Change-Id: I6bc63715b64ba67b1cb2348fadab16c9a41f61ba
JIRA: EE-1003

Former-commit-id: bb3d8363dba9951ad68da7953679655c80ad6bd9