Sanitize the CR_FeedbackFileManagingMock from data races
From time to time, we had abort signals while running the tests
due to this mock. When we run the Thread Sanitizer of Xcode,
the error is more explicit: the issue is related to a data race.

There is no @synchronize in Swift, we have to use a serial
queue or low level API. Here is a reference for the concurrency
approachs in swift:
http://www.cocoawithlove.com/blog/2016/06/02/threads-and-mutexes.html#conclusion

Change-Id: I39b46abafd2b831c5d3fc73a00678b895b7cd8d8
JIRA: EE-950

Former-commit-id: 252a857792bdee32e4964f7516a36708f224c1eb