Isolate tests from NSUserDefaults by mocking it in-memory
Using an in-memory NSUserDefaults isolates the tests.
With one of my simulators, I had red tests due to data
already stored in NSUserDefaults. It is important to
understand that [[NSUserDefaults alloc] init] doesn't
isolate the tests because the instance is backed by files.

An in-memory NSUserDefaults has the disadvantage to
be faster than the real one. So the integration tests
would be less representative than the real timing.
The advantage is that we can safely run tests in parallel
 while it is not possible with a classical NSUserDefaults.

Change-Id: Ic4e27bda8658ab3d71f74c3555200515c6bc6a77
JIRA: EE-NaN

Former-commit-id: 50e03943cc0986e0f45e8514c57d583fb4f0ac80