[tests] Split flaky tests
As it appears that some tests are flaky due to others, on factors that
are hard to control (WKWebView failure), let's try to isolate sensible
tests in a specific xcode target / test run
Note that this is named ITests as Integration Tests as it would probably
be good to isolate Integration tests first as they are the most
problematic ones. This is not fully done as splitting tests is not
trivial.

Change-Id: I87b813b038cc35214ea7c30501edef2008bbc61f
JIRA: EE-1028

Former-commit-id: abb92284532351da020ee592db052e42a1317e9f