Remove the Header Bidding logic from the BidManager
The goal is to separate the concerns. We extract all
the code related to the enrichment of the requests in a
dedicated class. "Header Bidding" as part of the
Domain Language didn't even exist in the code.

We should consider to inverse the dependency at
some point: CR_HeaderBidding would use the
CR_BidManager. Indeed, even if this logic
is encapsulated now, the CR_BidManager has too
many dependencies. To be done later of course.

Change-Id: Id25bdef48362a32fbc4446883318a5f58a929aef
JIRA: EE-1059

Former-commit-id: 37bffaf53103068d74450586e0aec763cb420d66