Refine NativeAdTableViewController to use multiple ads
The NativeAdLoader has a 1-n relation with the NativeAd.
In other term, we can have many loadAd for the same unit.
This change has for purpose to be able to test this
relation in the context of the impression (the loader
triggers the impression multiple time for diffent native
ad on the same adunit) and for the click.

Change-Id: If83f7ec47ac8d1103f5178c0d64493325c055f46
JIRA: EE-1093

Former-commit-id: 448b174428bcffeee15904874f6251692ef19af8