Put the dependency provider in Criteo instance
The bid manager builder holds more than what the bid manager needs. And
this prevented other components, above or transverse to the bid manager
(such as CSM, bidding integrations), to access dependencies. A
workaround was to ask the criteo instance which ask the bid manager for
the dependencies.

This creates a god object of dependencies that has no behaviour, except
creating components. It is hold at the highest level: the Criteo
instance.

Also, from the tests, this allows to mock specific part of the behaviour
while building the other as in prod code. So instantiation may be
factorized in this.

Change-Id: Iaf6946046b1bb21ffe5ed37ea2ebbe84f64ad477
JIRA: EE-1060

Former-commit-id: 4b71e2a67f2605e729a23d60de795d39876beafd