Refine utils methods in the integration tests of the native ad
Concretely, we inspect dynamically the given parameter in the
modified methods of this commit. The type of the parameter isn't
really useful and it even block us for reusing the logic with the both
UIViewController types that we have. So we use a more
general type and assert on the availability of the property
on the object.

The 'expectationForImpressionOnViewController:...' will
be used in next commit with the CR_NativeAdViewController.

Change-Id: I51829c15d6823dff5d38d8fce82bcb27e40c71d8
JIRA: EE-1093

Former-commit-id: 1a3cffe762e1dc9410c1f5be962ccee755513f3a