[DeviceInfo] Fix WebView init deadlock
Quoting apple documentation: https://developer.apple.com/library/archive/documentation/General/Conceptual/ConcurrencyProgrammingGuide/OperationQueues/OperationQueues.html#//apple_ref/doc/uid/TP40008091-CH102-SW28
> Do not call the dispatch_sync function from a task that is executing
> on the same queue that you pass to your function call. Doing so will
> deadlock the queue. If you need to dispatch to the current queue, do
> so asynchronously using the dispatch_async function.

Before fix, we do a dispatch_sync if not on the main thread.
Main Queue targets a specific threadpool, which includes the main
thread, but not only, as a result that could deadlock if we were on
main queue but not main thread. As it is quite complicated to check if
we are on the main queue, let's just lazily instantiate the webview on
main queue and clearly name / hide the constructor that is for testing
only.

Change-Id: Ib05f9102430f2254baa335f70b47c89e5b1bcdae

Former-commit-id: d69e9cc57adafee1b3a374ab4a2aa1ce6520c498