[Actions] Java & WireMock setup
Note that WireMock setup is a bit hack-ish as it downloads the jar and
starts it as a background process. A better approach would be to use
a github action to do this, to use toolcache and properly start and
stop WireMock. There is [setup-wiremock] but it does not allow
starting with https.

[setup-wiremock]: https://github.com/marketplace/actions/setup-wiremock