[BidManager] Fix CSM tests
Tests were broken because of mocking NSDate to `TimeIntervalSince1970:0`.
As the consequence the bidmanager was in user silent mode, not fetching
a bid after a consume, and not sending CSM while it shoud have.
The issue appeared after refactoring the user silent mode but was there
before that.
These tests will probably be reworked in upcoming changes for CSM with
live bidding.