[Native] Keep loader alive with ad
That way client code does not need to keep a reference to the loader.
The loader will be alive until the ad and related view will be in
memory.
Side note: that reinforce the fact a loader is coupled to an ad load
and display, and is not well adapted to load several ads as it is hard
to determine from delegate calls which ad was impressed or clicked.

JIRA: EE-1209